# [Xamarin.Android.Build.Tasks] Restore ConvertResourcesCases skipping