@brendanzagaeski 2019-06-22
This little build time increase is present in the current previews from the `d16-2` branch compared to the current release from the `d16-1` branch, so it might be a candidate to cherry-pick.

I noticed this behavior by chance while I was making sure I understood what was happening on <https://github.com/xamarin/xamarin-android/issues/3250>, but that other issue is not directly related to this `<ConvertResourcesCases/>` behavior.