@jonathanpeppers 2019-06-22
Crap I thought I saw this in a build log when testing AndroidX, but I didn’t look into it...