@jonathanpeppers 2019-06-24
Yeah latest perf tests show this caused:

![image](https://user-images.githubusercontent.com/840039/60031458-69f6ea00-966a-11e9-8ac8-b70bec41199e.png)